Implement DLL project for CUPS API on WIN32 (STR #366)
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@4828 7a7537e8-13f0-0310-91df-b6672ffda945

+++ cups/http.h
@@ -445,7 +445,7 @@ extern http_uri_status_t httpSeparateURI(const char *uri,
 			        	 char *username, int usernamelen,
 					 char *host, int hostlen, int *port,
 					 char *resource, int resourcelen);
-extern void		httpSetLength(http_t *http, off_t length);
+extern void		httpSetLength(http_t *http, size_t length);
 
 
 /*
