+++ cups/md5passwd.c
@@ -60,7 +60,7 @@ httpMD5(const char *username,		/* I - User name */
 
   snprintf(line, sizeof(line), "%s:%s:%s", username, realm, passwd);
   _cups_md5_init(&state);
-  _cups_md5_append(&state, (unsigned char *)line, strlen(line));
+  _cups_md5_append(&state, (unsigned char *)line, (int)strlen(line));
   _cups_md5_finish(&state, sum);
 
  /*
@@ -95,7 +95,7 @@ httpMD5Final(const char *nonce,		/* I - Server nonce value */
 
   snprintf(line, sizeof(line), "%s:%s", method, resource);
   _cups_md5_init(&state);
-  _cups_md5_append(&state, (unsigned char *)line, strlen(line));
+  _cups_md5_append(&state, (unsigned char *)line, (int)strlen(line));
   _cups_md5_finish(&state, sum);
   httpMD5String(sum, a2);
 
@@ -108,7 +108,7 @@ httpMD5Final(const char *nonce,		/* I - Server nonce value */
   snprintf(line, sizeof(line), "%s:%s:%s", md5, nonce, a2);
 
   _cups_md5_init(&state);
-  _cups_md5_append(&state, (unsigned char *)line, strlen(line));
+  _cups_md5_append(&state, (unsigned char *)line, (int)strlen(line));
   _cups_md5_finish(&state, sum);
 
   return (httpMD5String(sum, md5));
